requirements.txt
langgraph==1.0.5
langchain-openai==1.1.3
fastapi==0.124.4
uvicorn==0.38.0
httpx==0.28.1
pydantic==2.12.5
python-dotenv==1.2.1
pytest==9.0.2
pytest-asyncio==1.3.0


PyJWT==2.8.0
